backend/requirements.txt
Flask==3.1.0
Flask-CORS==5.0.0
onnx==1.20.1
onnxruntime==1.24.1
transformers==4.45.2
numpy==1.26.4
google-api-python-client==2.151.0
python-dotenv==1.0.1